cc2/requirements.txt
asgiref==3.2.7
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
Django==3.0.5
gcloud==0.18.3
google-api-core==1.17.0
google-auth==1.14.2
google-cloud==0.34.0
google-cloud-core==1.3.0
google-cloud-storage==1.28.1
google-cloud-vision==1.0.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
grpcio==1.28.1
httplib2==0.17.3
idna==2.9
mysqlclient==1.4.6
oauth2client==4.1.3
protobuf==3.11.3
psycopg2==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2019.3
requests==2.23.0
rsa==4.0
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.9
vision==1.0.0
